fallback: module prototype
The module switches individual queries to a forwarding config
when failing to resolve the normal way (non-forwarding typically).
Internally the mechanism is close to the serve_stale module.

For now it has Lua config only, e.g.:

modules = { 'fallback' }
fallback.config({
	targets = {
		{'193.17.47.1', tls = false},
		{'2001:148f:fffe::1', tls = false},
	},
	options = {
	},
})


TODO: right now it's not yet possible to specify a list of source
addresses for which this mechanism won't be applied.